fix: 최적화

+++ pages/api/post/delete.ts
@@ -1,17 +1,27 @@
+import { authOptions } from '@/pages/api/auth/[...nextauth]';
 import { connectDB } from '@/util/database';
 import { ObjectId } from 'mongodb';
+import { getServerSession } from 'next-auth';
 
 export default async function handler(req: any, res: any) {
     if (req.method === 'DELETE') {
+        const session = await getServerSession(req, res, authOptions);
+
         try {
             const db = (await connectDB).db('forum');
-            const result = await db
+
+            const find = await db
                 .collection('post')
-                .deleteOne({ _id: new ObjectId(req.body) });
+                .findOne({ _id: new ObjectId(req.body) });
 
-            if (result.deletedCount < 1) res.status(500).json('삭제 실패임');
-
-            res.status(200).json('삭제완료');
+            if (find?.author === session?.user?.email) {
+                const result = await db
+                    .collection('post')
+                    .deleteOne({ _id: new ObjectId(req.body) });
+                res.status(200).json('삭제완료');
+            } else {
+                res.status(500).json('님이 쓴 글 아님');
+            }
         } catch (error) {
             throw error;
         }
